refactored logic
