+++ digitalpy/core/logic/__init__.py
@@ -1,0 +1,2 @@
+"""describes the abstract strategy to business logic and it's Implementation
+"""
