+++ digitalpy/core/logic/business_rule_controller.py
@@ -1,4 +1,4 @@
-from digitalpy.routing.controller import Controller
+from digitalpy.core.zmanager.controller import Controller
 from abc import ABC, abstractmethod
 from typing import Union
